Tag and release 2.2.1
git-svn-id: https://plugins.svn.wordpress.org/backupwordpress/tags/2.2.1@667431 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ admin/actions.php
@@ -9,8 +9,9 @@ function hmbkp_request_delete_backup() {
 	if ( empty( $_GET['hmbkp_delete_backup'] ) || ! check_admin_referer( 'hmbkp-delete_backup' ) )
 		return;
 
-	$schedule = new HMBKP_Scheduled_Backup( urldecode( $_GET['hmbkp_schedule_id'] ) );
-	$schedule->delete_backup( base64_decode( urldecode( $_GET['hmbkp_delete_backup'] ) ) );
+	$schedule = new HMBKP_Scheduled_Backup( sanitize_text_field( urldecode( $_GET['hmbkp_schedule_id'] ) ) );
+
+	$schedule->delete_backup( sanitize_text_field( base64_decode( $_GET['hmbkp_delete_backup'] ) ) );
 
 	wp_redirect( remove_query_arg( array( 'hmbkp_delete_backup', '_wpnonce' ) ), 303 );
 
@@ -28,7 +29,7 @@ function hmbkp_request_delete_schedule() {
 	if ( empty( $_GET['action'] ) || $_GET['action'] !== 'hmbkp_delete_schedule' || ! check_admin_referer( 'hmbkp-delete_schedule' ) )
 		return;
 
-	$schedule = new HMBKP_Scheduled_Backup( urldecode( $_GET['hmbkp_schedule_id'] ) );
+	$schedule = new HMBKP_Scheduled_Backup( sanitize_text_field( urldecode( $_GET['hmbkp_schedule_id'] ) ) );
 	$schedule->cancel( true );
 
 	wp_redirect( remove_query_arg( array( 'hmbkp_schedule_id', 'action', '_wpnonce' ) ), 303 );
@@ -43,18 +44,40 @@ add_action( 'load-tools_page_' . HMBKP_PLUGIN_SLUG, 'hmbkp_request_delete_schedu
  */
 function hmbkp_ajax_request_do_backup() {
 
-	if ( empty( $_GET['hmbkp_schedule_id'] ) )
-		return;
+	if ( empty( $_POST['hmbkp_schedule_id'] ) )
+		exit;
+
+	//header( 'HTTP/1.1 500 Internal Server Error' );
+
+	// We want to display any fatal errors in this ajax request so we can catch them on the other side.
+	error_reporting( E_ERROR | E_USER_ERROR | E_CORE_ERROR | E_COMPILE_ERROR | E_RECOVERABLE_ERROR );
+	@ini_set( 'display_errors', 'On' );
+	@ini_set( 'html_errors', 'Off' );
+
+	// Force a memory error for testing purposes
+	//ini_set( 'memory_limit', '2M' );
+	//function a() { a(); } a();
 
 	ignore_user_abort( true );
 
 	hmbkp_cleanup();
 
-	$schedule = new HMBKP_Scheduled_Backup( urldecode( $_GET['hmbkp_schedule_id'] ) );
+	$schedule = new HMBKP_Scheduled_Backup( sanitize_text_field( urldecode( $_POST['hmbkp_schedule_id'] ) ) );
 
 	$schedule->run();
 
-	hmbkp_schedule_actions( $schedule );
+	$errors = array_merge( $schedule->get_errors(), $schedule->get_warnings() );
+
+	$error_message = '';
+
+	foreach ( $errors as $error_set )
+		$error_message .= implode( "\n\r", $error_set );
+
+	if ( $error_message && file_exists( $schedule->get_archive_filepath() ) )
+		$error_message .= 'HMBKP_SUCCESS';
+
+	if ( trim( $error_message ) )
+		echo $error_message;
 
 	exit;
 
@@ -67,10 +90,26 @@ add_action( 'wp_ajax_hmbkp_run_schedule', 'hmbkp_ajax_request_do_backup' );
  */
 function hmbkp_request_download_backup() {
 
-	if ( empty( $_GET['hmbkp_download_backup'] ) || ! check_admin_referer( 'hmbkp-download_backup' ) )
+	global $is_apache;
+
+	if ( empty( $_GET['hmbkp_download_backup'] ) || ! check_admin_referer( 'hmbkp-download_backup' ) || ! file_exists( sanitize_text_field( base64_decode( $_GET['hmbkp_download_backup'] ) ) ) )
 		return;
 
-	wp_redirect( str_replace( realpath( HM_Backup::conform_dir( HM_Backup::get_home_path() ) ), home_url(), base64_decode( $_GET['hmbkp_download_backup'] ) ), 303 );
+	$url = str_replace( HM_Backup::conform_dir( HM_Backup::get_home_path() ), home_url(), trailingslashit( dirname( sanitize_text_field( base64_decode( $_GET['hmbkp_download_backup'] ) ) ) ) ) . urlencode( pathinfo( sanitize_text_field( base64_decode( $_GET['hmbkp_download_backup'] ) ), PATHINFO_BASENAME ) );
+
+	if ( $is_apache ) {
+
+		// Force the .htaccess to be rebuilt
+		if ( file_exists( hmbkp_path() . '/.htaccess' ) )
+			unlink( hmbkp_path() . '/.htaccess' );
+
+		hmbkp_path();
+
+		$url = add_query_arg( 'key', HMBKP_SECURE_KEY, $url );
+
+	}
+
+	wp_redirect( $url, 303 );
 
 	exit;
 
@@ -86,10 +125,10 @@ function hmbkp_request_cancel_backup() {
 	if ( ! isset( $_GET['action'] ) || $_GET['action'] !== 'hmbkp_cancel' )
 		return;
 
-	$schedule = new HMBKP_Scheduled_Backup( urldecode( $_GET['hmbkp_schedule_id'] ) );
+	$schedule = new HMBKP_Scheduled_Backup( sanitize_text_field( urldecode( $_GET['hmbkp_schedule_id'] ) ) );
 
 	// Delete the running backup
-	if ( file_exists( trailingslashit( hmbkp_path() ) . $schedule->get_running_backup_filename() ) )
+	if ( $schedule->get_running_backup_filename() && file_exists( trailingslashit( hmbkp_path() ) . $schedule->get_running_backup_filename() ) )
 		unlink( trailingslashit( hmbkp_path() ) . $schedule->get_running_backup_filename() );
 
 	hmbkp_cleanup();
@@ -124,10 +163,10 @@ add_action( 'admin_init', 'hmbkp_dismiss_error' );
  */
 function hmbkp_ajax_is_backup_in_progress() {
 
-	if ( empty( $_GET['hmbkp_schedule_id'] ) )
-		return;
+	if ( empty( $_POST['hmbkp_schedule_id'] ) )
+		exit;
 
-	$schedule = new HMBKP_Scheduled_Backup( urldecode( $_GET['hmbkp_schedule_id'] ) );
+	$schedule = new HMBKP_Scheduled_Backup( sanitize_text_field( urldecode( $_POST['hmbkp_schedule_id'] ) ) );
 
 	if ( ! $schedule->get_status() )
 		echo 0;
@@ -145,10 +184,10 @@ add_action( 'wp_ajax_hmbkp_is_in_progress', 'hmbkp_ajax_is_backup_in_progress' )
  */
 function hmbkp_ajax_calculate_backup_size() {
 
-	if ( empty( $_GET['hmbkp_schedule_id'] ) )
-		return;
+	if ( empty( $_POST['hmbkp_schedule_id'] ) )
+		exit;
 
-	$schedule = new HMBKP_Scheduled_Backup( urldecode( $_GET['hmbkp_schedule_id'] ) );
+	$schedule = new HMBKP_Scheduled_Backup( sanitize_text_field( urldecode( $_POST['hmbkp_schedule_id'] ) ) );
 
 	$recalculate_filesize = true;
 
@@ -164,10 +203,13 @@ add_action( 'wp_ajax_hmbkp_calculate', 'hmbkp_ajax_calculate_backup_size' );
  */
 function hmbkp_ajax_cron_test() {
 
-	$response = wp_remote_get( site_url( 'wp-cron.php' ) );
+	$response = wp_remote_head( site_url( 'wp-cron.php' ), array( 'timeout' => 30 ) );
 
-	if ( ! is_wp_error( $response ) && $response['response']['code'] != '200' )
-    	echo '<div id="hmbkp-warning" class="updated fade"><p><strong>' . __( 'BackUpWordPress has detected a problem.', 'hmbkp' ) . '</strong> ' . sprintf( __( '%1$s is returning a %2$s response which could mean cron jobs aren\'t getting fired properly. BackUpWordPress relies on wp-cron to run scheduled back ups. See the %3$s for more details.', 'hmbkp' ), '<code>wp-cron.php</code>', '<code>' . $response['response']['code'] . '</code>', '<a href="http://wordpress.org/extend/plugins/backupwordpress/faq/">FAQ</a>' ) . '</p></div>';
+	if ( is_wp_error( $response ) )
+		echo '<div id="hmbkp-warning" class="updated fade"><p><strong>' . __( 'BackUpWordPress has detected a problem.', 'hmbkp' ) . '</strong> ' . sprintf( __( '%1$s is returning a %2$s response which could mean cron jobs aren\'t getting fired properly. BackUpWordPress relies on wp-cron to run scheduled back ups. See the %3$s for more details.', 'hmbkp' ), '<code>wp-cron.php</code>', '<code>' . $response->get_error_message() . '</code>', '<a href="http://wordpress.org/extend/plugins/backupwordpress/faq/">FAQ</a>' ) . '</p></div>';
+
+	elseif ( wp_remote_retrieve_response_code( $response ) != 200 )
+    	echo '<div id="hmbkp-warning" class="updated fade"><p><strong>' . __( 'BackUpWordPress has detected a problem.', 'hmbkp' ) . '</strong> ' . sprintf( __( '%1$s is returning a %2$s response which could mean cron jobs aren\'t getting fired properly. BackUpWordPress relies on wp-cron to run scheduled back ups. See the %3$s for more details.', 'hmbkp' ), '<code>wp-cron.php</code>', '<code>' . esc_html( wp_remote_retrieve_response_code( $response ) ) . ' ' . esc_html( get_status_header_desc( wp_remote_retrieve_response_code( $response ) ) ) . '</code>', '<a href="http://wordpress.org/extend/plugins/backupwordpress/faq/">FAQ</a>' ) . '</p></div>';
 
 	else
 		echo 1;
@@ -182,7 +224,10 @@ add_action( 'wp_ajax_hmbkp_cron_test', 'hmbkp_ajax_cron_test' );
  */
 function hmbkp_edit_schedule_load() {
 
-	$schedule = new HMBKP_Scheduled_Backup( esc_attr( $_GET['hmbkp_schedule_id'] ) );
+	if ( empty( $_GET['hmbkp_schedule_id'] ) )
+		exit;
+
+	$schedule = new HMBKP_Scheduled_Backup( sanitize_text_field( $_GET['hmbkp_schedule_id'] ) );
 
 	require( HMBKP_PLUGIN_PATH . '/admin/schedule-form.php' );
 
@@ -196,7 +241,10 @@ add_action( 'wp_ajax_hmbkp_edit_schedule_load', 'hmbkp_edit_schedule_load' );
  */
 function hmbkp_edit_schedule_excludes_load() {
 
-	$schedule = new HMBKP_Scheduled_Backup( esc_attr( $_GET['hmbkp_schedule_id'] ) );
+	if ( empty( $_GET['hmbkp_schedule_id'] ) )
+		exit;
+
+	$schedule = new HMBKP_Scheduled_Backup( sanitize_text_field( $_GET['hmbkp_schedule_id'] ) );
 
 	require( HMBKP_PLUGIN_PATH . '/admin/schedule-form-excludes.php' );
 
@@ -228,56 +276,66 @@ add_action( 'wp_ajax_hmbkp_add_schedule_load', 'hmbkp_add_schedule_load' );
 function hmnkp_edit_schedule_submit() {
 
 	if ( empty( $_GET['hmbkp_schedule_id'] ) )
-		return;
+		exit;
 
-	$schedule = new HMBKP_Scheduled_Backup( esc_attr( $_GET['hmbkp_schedule_id'] ) );
+	$schedule = new HMBKP_Scheduled_Backup( sanitize_text_field( $_GET['hmbkp_schedule_id'] ) );
 
 	$errors = array();
 
+	
+
 	if ( isset( $_GET['hmbkp_schedule_type'] ) ) {
 
-		if ( ! trim( $_GET['hmbkp_schedule_type'] ) )
+		$schedule_type = sanitize_text_field( $_GET['hmbkp_schedule_type'] );
+
+		if ( ! trim( $schedule_type ) )
 			$errors['hmbkp_schedule_type'] = __( 'Backup type cannot be empty', 'hmbkp' );
 
-		elseif ( ! in_array( $_GET['hmbkp_schedule_type'], array( 'complete', 'file', 'database' ) ) )
+		elseif ( ! in_array( $schedule_type, array( 'complete', 'file', 'database' ) ) )
 			$errors['hmbkp_schedule_type'] = __( 'Invalid backup type', 'hmbkp' );
 
 		else
-			$schedule->set_type( $_GET['hmbkp_schedule_type'] );
+			$schedule->set_type( $schedule_type );
 
 	}
 
 	if ( isset( $_GET['hmbkp_schedule_reoccurrence'] ) ) {
 
-		if ( empty( $_GET['hmbkp_schedule_reoccurrence'] ) )
+		$schedule_reoccurrence = sanitize_text_field( $_GET['hmbkp_schedule_reoccurrence'] );
+
+		if ( empty( $schedule_reoccurrence ) )
 			$errors['hmbkp_schedule_reoccurrence'] = __( 'Schedule cannot be empty', 'hmbkp' );
 
-		elseif ( ! in_array( $_GET['hmbkp_schedule_reoccurrence'], array_keys( wp_get_schedules() ) ) && $_GET['hmbkp_schedule_reoccurrence'] !== 'manually' )
+		elseif ( ! in_array( $schedule_reoccurrence, array_keys( hmbkp_cron_schedules() ) ) && $schedule_reoccurrence !== 'manually' )
 			$errors['hmbkp_schedule_reoccurrence'] = __( 'Invalid schedule', 'hmbkp' );
 
 		else
-			$schedule->set_reoccurrence( $_GET['hmbkp_schedule_reoccurrence'] );
+			$schedule->set_reoccurrence( $schedule_reoccurrence );
 
 	}
 
 	if ( isset( $_GET['hmbkp_schedule_max_backups'] ) ) {
 
-		if ( empty( $_GET['hmbkp_schedule_max_backups'] ) )
+		$schedule_max_backups = sanitize_text_field( $_GET['hmbkp_schedule_max_backups'] );
+
+		if ( empty( $schedule_max_backups ) )
 			$errors['hmbkp_schedule_max_backups'] = __( 'Max backups can\'t be empty', 'hmbkp' );
 
-		elseif ( ! is_numeric( $_GET['hmbkp_schedule_max_backups'] ) )
+		elseif ( ! is_numeric( $schedule_max_backups ) )
 			$errors['hmbkp_schedule_max_backups'] = __( 'Max backups must be a number', 'hmbkp' );
 
-		elseif ( ! ( $_GET['hmbkp_schedule_max_backups'] >= 1 ) )
+		elseif ( ! ( $schedule_max_backups >= 1 ) )
 			$errors['hmbkp_schedule_max_backups'] = __( 'Max backups must be greater than 0', 'hmbkp' );
 
 		else
-			$schedule->set_max_backups( (int) $_GET['hmbkp_schedule_max_backups'] );
+			$schedule->set_max_backups( (int) $schedule_max_backups );
 
+		// Remove any old backups in-case max backups was reduced
 		$schedule->delete_old_backups();
 
 	}
 
+	// Save the service options
 	foreach ( HMBKP_Services::get_services( $schedule ) as $service )
         $errors = array_merge( $errors, $service->save() );
 
@@ -300,13 +358,16 @@ add_action( 'wp_ajax_hmnkp_edit_schedule_submit', 'hmnkp_edit_schedule_submit' )
  */
 function hmbkp_add_exclude_rule() {
 
-	$schedule = new HMBKP_Scheduled_Backup( esc_attr( $_POST['hmbkp_schedule_id'] ) );
+	if ( empty( $_POST['hmbkp_schedule_id'] ) )
+		exit;
 
-	$schedule->set_excludes( $_POST['hmbkp_exclude_rule'], true );
+	$schedule = new HMBKP_Scheduled_Backup( sanitize_text_field( $_POST['hmbkp_schedule_id'] ) );
+
+	$schedule->set_excludes( sanitize_text_field( $_POST['hmbkp_exclude_rule'] ), true );
 
 	$schedule->save();
 
-	include( HMBKP_PLUGIN_PATH . '/admin/schedule-form-excludes.php' );
+	require( HMBKP_PLUGIN_PATH . '/admin/schedule-form-excludes.php' );
 
 	exit;
 
@@ -322,15 +383,18 @@ add_action( 'wp_ajax_hmbkp_add_exclude_rule', 'hmbkp_add_exclude_rule' );
  */
 function hmbkp_delete_exclude_rule() {
 
-	$schedule = new HMBKP_Scheduled_Backup( esc_attr( $_POST['hmbkp_schedule_id'] ) );
+	if ( empty( $_GET['hmbkp_schedule_id'] ) )
+		exit;
+
+	$schedule = new HMBKP_Scheduled_Backup( sanitize_text_field( $_GET['hmbkp_schedule_id'] ) );
 
 	$excludes = $schedule->get_excludes();
 
-	$schedule->set_excludes( array_diff( $excludes, (array) $_POST['hmbkp_exclude_rule'] ) );
+	$schedule->set_excludes( array_diff( $excludes, (array) stripslashes( sanitize_text_field( $_GET['hmbkp_exclude_rule'] ) ) ) );
 
 	$schedule->save();
 
-	include( HMBKP_PLUGIN_PATH . '/admin/schedule-form-excludes.php' );
+	require( HMBKP_PLUGIN_PATH . '/admin/schedule-form-excludes.php' );
 
 	exit;
 
@@ -346,55 +410,76 @@ add_action( 'wp_ajax_hmbkp_delete_exclude_rule', 'hmbkp_delete_exclude_rule' );
  */
 function hmbkp_preview_exclude_rule() {
 
-	if ( ! empty( $_POST['hmbkp_schedule_id'] ) )
-		$schedule = new HMBKP_Scheduled_Backup( $_POST['hmbkp_schedule_id'] );
+	if ( empty( $_POST['hmbkp_schedule_id'] ) || empty( $_POST['hmbkp_schedule_excludes'] ) )
+		exit;
 
-	if ( ! empty( $_POST['hmbkp_schedule_excludes'] ) )
-		$excludes = explode( ',', $_POST['hmbkp_schedule_excludes'] );
+	$schedule = new HMBKP_Scheduled_Backup( sanitize_text_field( $_POST['hmbkp_schedule_id'] ) );
 
-	if ( ! empty( $_POST['hmbkp_file_method'] ) )
-		$file_method = $_POST['hmbkp_file_method'];
+	$excludes = explode( ',', sanitize_text_field( $_POST['hmbkp_schedule_excludes'] ) );
 
-	hmbkp_file_list( $schedule, $excludes, $file_method );
+	hmbkp_file_list( $schedule, $excludes, 'get_excluded_files' );
 
-	foreach( $schedule->get_excluded_files() as $key => $excluded_file )
-		if ( strpos( $excluded_file, $schedule->get_path() ) === false )
-			$excluded_files[] = $excluded_file;
+	$schedule->set_excludes( $excludes );
 
-	if ( ! empty( $excluded_files) ) { ?>
+	if ( $schedule->get_excluded_file_count() ) { ?>
 
-	<p><?php printf( _n( '%s matches 1 file.', '%1$s matches %2$d files.', count( $excluded_files ), 'hmbkp' ), '<code>' . implode( '</code>, <code>', $excludes ) . '</code>', count( $excluded_files ) ); ?></p>
+		<p><?php printf( _n( '%s matches 1 file.', '%1$s matches %2$d files.', $schedule->get_excluded_file_count(), 'hmbkp' ), '<code>' . implode( '</code>, <code>', array_map( 'esc_html', $excludes ) ) . '</code>', $schedule->get_excluded_file_count() ); ?></p>
 
 	<?php } else { ?>
 
-	<p><?php printf( __( '%s didn\'t match any files.', 'hmbkp' ), '<code>' . implode( '</code>, <code>', $excludes ) . '</code>' ); ?></p>
+		<p><?php printf( __( '%s didn\'t match any files.', 'hmbkp' ), '<code>' . implode( '</code>, <code>', array_map( 'esc_html', $excludes ) ) . '</code>' ); ?></p>
 
 	<?php } ?>
 
-	<p><button type="button" class="button-primary hmbkp_save_exclude_rule"><?php _e( 'Exclude', 'hmbkp' ); ?></button> <button type="button" class="button-secondary hmbkp_cancel_save_exclude_rule"><?php _e( 'Cancel', 'hmbkp' ); ?></button></p>
+		<p><button type="button" class="button-primary hmbkp_save_exclude_rule"><?php _e( 'Exclude', 'hmbkp' ); ?></button> <button type="button" class="button-secondary hmbkp_cancel_save_exclude_rule"><?php _e( 'Cancel', 'hmbkp' ); ?></button></p>
 
 	<?php exit;
 
 }
 add_action( 'wp_ajax_hmbkp_file_list', 'hmbkp_preview_exclude_rule', 10, 0 );
 
-/**
- * Handles changes in the defined Constants
- * that users can define to control advanced
- * settings
- */
-function hmbkp_constant_changes() {
+function hmbkp_display_error_and_offer_to_email_it() {
 
-	// If a custom backup path has been set or changed
-	if ( defined( 'HMBKP_PATH' ) && HMBKP_PATH && HM_Backup::conform_dir( HMBKP_PATH ) !== ( $from = HM_Backup::conform_dir( get_option( 'hmbkp_path' ) ) ) )
-		hmbkp_path_move( $from, HMBKP_PATH );
+	if ( empty( $_POST['hmbkp_error'] ) )
+		exit;
 
-	// If a custom backup path has been removed
-	if ( ( ( defined( 'HMBKP_PATH' ) && ! HMBKP_PATH ) || ! defined( 'HMBKP_PATH' ) && HM_Backup::conform_dir( hmbkp_path_default() ) != ( $from = HM_Backup::conform_dir( get_option( 'hmbkp_path' ) ) ) ) )
-		hmbkp_path_move( $from, hmbkp_path_default() );
+	$error = wp_strip_all_tags( stripslashes( $_POST['hmbkp_error'] ) ); 
 
-	// If the custom path has changed and the new directory isn't writable
-	if ( defined( 'HMBKP_PATH' ) && HMBKP_PATH && HM_Backup::conform_dir( HMBKP_PATH ) != ( $from = HM_Backup::conform_dir( get_option( 'hmbkp_path' ) ) ) && $from != hmbkp_path_default() && !is_writable( HMBKP_PATH ) && is_dir( $from ) )
-		hmbkp_path_move( $from, hmbkp_path_default() );
+	$error = str_replace( 'HMBKP_SUCCESS', '', $error, $succeeded ); 
+
+	if ( $succeeded ) { ?>
+
+		<h3><?php _e( 'Your backup completed but with the following errors / warnings', 'hmbkp' ); ?></h3>
+
+	<?php } else { ?>
+
+		<h3><?php _e( 'Your backUp failed', 'hmbkp' ); ?></h3>
+
+	<?php } ?>
+
+	<p><?php _e( 'Here\'s the response from the server:', 'hmbkp' ); ?></p>
+
+	<pre><?php esc_html_e( $error ); ?></pre>
+
+	<p class="description"><?php printf( __( 'You can email details of this error to %s so they can look into the issue.', 'hmbkp' ), '<a href="http://hmn.md">Human Made Limited</a>' ); ?><br /><br /></p>
+
+	<button class="button hmbkp-fancybox-close"><?php _e( 'Close', 'hmbkp' ); ?></button>
+	<button class="button-primary hmbkp_send_error_via_email right"><?php _e( 'Email to Support', 'hmbkp' ); ?></button>
+
+	<?php exit;
+}
+add_action( 'wp_ajax_hmbkp_backup_error', 'hmbkp_display_error_and_offer_to_email_it' );
+
+function hmbkp_send_error_via_email() {
+
+	if ( empty( $_POST['hmbkp_error'] ) )
+		exit;
+
+	$error = wp_strip_all_tags( $_POST['hmbkp_error'] );
+
+	wp_mail( 'support@humanmade.co.uk', 'BackUpWordPress Fatal error on ' . parse_url( home_url(), PHP_URL_HOST ), $error, 'From: BackUpWordPress <' . get_bloginfo( 'admin_email' ) . '>' . "\r\n" );
+
+	exit;
 
 }
+add_action( 'wp_ajax_hmbkp_email_error', 'hmbkp_send_error_via_email' );
